Consolidate multiple Mii test classes to a single ItClass  (#1875)
* Consolidate Mii Domains and remove junit4 tests

* removed more Junit4 Mii tests

* Modify the logic to check SystemResources

* Updated the initAll() with  installAndVerifyOperator replacing the old code

* Modify the assertion for delete resources
