+++ new-integration-tests/src/test/resources/wdt-models/model.delete.sysresources.yaml
@@ -1,0 +1,15 @@
+# Copyright (c) 2020, Oracle Corporation and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+resources:
+    FileStore:
+        !TestClusterFileStore:
+    JMSServer:
+        !TestClusterJmsServer:
+    JMSSystemResource:
+        !TestClusterJmsModule:
+    JDBCSystemResource:
+        !TestDataSource:
+appDeployments:
+    Application:
+        !myear:
