+++ new-integration-tests/src/test/resources/wdt-models/model.jms2.yaml
@@ -2,30 +2,30 @@
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 resources:
     FileStore:
-        TestClusterFileStore:
+        TestClusterFileStore2:
            Target: 'cluster-1'
     JMSServer:
-        TestClusterJmsServer:
+        TestClusterJmsServer2:
             ProductionPausedAtStartup: false
             ConsumptionPausedAtStartup: false
             Target: 'cluster-1'
-            PersistentStore: 'TestClusterFileStore'
+            PersistentStore: 'TestClusterFileStore2'
             InsertionPausedAtStartup: false
             MessageCompressionOptions: GZIP_DEFAULT_COMPRESSION
 
     JMSSystemResource:
-        TestClusterJmsModule:
+        TestClusterJmsModule2:
             Target: 'cluster-1'
             SubDeployment:
-                TestClusterSubDeployment:
-                    Target: TestClusterJmsServer
+                TestClusterSubDeployment2:
+                    Target: TestClusterJmsServer2
             JmsResource:
               UniformDistributedQueue:
-                testUniformQueue:
-                   SubDeploymentName: TestClusterSubDeployment
-                   JNDIName: jms/testUniformQueue
+                testUniformQueue2:
+                   SubDeploymentName: TestClusterSubDeployment2
+                   JNDIName: jms/testUniformQueue2
               UniformDistributedTopic:
-                testUniformTopic:
-                   SubDeploymentName: TestClusterSubDeployment
+                testUniformTopic2:
+                   SubDeploymentName: TestClusterSubDeployment2
                    ForwardingPolicy: Partitioned
-                   JNDIName: jms/testUniformTopic
+                   JNDIName: jms/testUniformTopic2
