--- new-integration-tests/src/test/resources/wdt-models/model.wldf.yaml
@@ -1,8 +1,0 @@
-# Copyright (c) 2020, Oracle Corporation and/or its affiliates.
-# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-resources:
-    WLDFSystemResource:
-        'TestWldfModule':
-            Description: WLDF Module for Situational Config Override
-            Target: 'cluster-1'
-
